extensive changes..
am seperating out in dreamsock folder and basegame folder all the classes...
what i mean is...
I am getting rid of structs and turning them into classes. I am
getting rid of multiple classes in one file.
I am renaming files to coincide with the class they contain. I am not
at this point trying to get anything to work....
I believe in doing this I will understand what exactly is going on in
the code and then it can more easily be read by us
and modified and extended.

this does not compile and is work in progress.